backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-countries==7.6.1
django-environ==0.12.0
django-phonenumber-field==8.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
